# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Proyecto Express API",
    "description": "Documentación generada automáticamente con swagger-autogen",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://62.169.28.169"
    }
  ],
  "paths": {
...
    "./genre-Pel/{genre}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "genre",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}